X-Code Mac 8c45e287f3ac58565fe32924ea68b17f
8932c6a96198670d480e00078c46d3c04b7ec4eb456aa36da8cd95f3f0056a13

+++ Code Template/Code_Template/Headers/functions.h
@@ -167,7 +167,7 @@ static void sizes()
     std::cout << "long long int : " <<  8 * sizeof(long long int) << " bits" <<  std::endl;
 }
 
-long double unsigned_ripple_carry_adder(uint64_t a, uint64_t b)
+uint64_t unsigned_ripple_carry_adder(uint64_t a, uint64_t b)
 {
     uint64_t carryin = 0, sum = 0, mask = 1, temp_a = a, temp_b = b;
     while(temp_a || temp_b)
